+++ src/app/admin/lunch/page.tsx
@@ -15,7 +15,7 @@ export default function LunchAdmin() {
     return (
         <OnlyRoles roles={["administrator"]}>
             <PageWithHeader title="Admin / Ebédrendelés" homeLocation="/admin">
-                <div className="flex flex-col justify-center lg:flex-row">
+                <div className="flex flex-col justify-center xl:flex-row">
                     <Card>
                         <SetMenuAndSendEmail />
                     </Card>