+++ src/models/User.model.ts
@@ -9,6 +9,8 @@ export interface IUser {
     groups: Types.ObjectId[];
     nfcId: string;
     laptopPasswordChanged?: Date;
+    autoOrder?: { chosen: string }[];
+    blocked?: boolean;
 }
 
 const userSchema = new Schema<IUser>({
@@ -35,6 +37,18 @@ const userSchema = new Schema<IUser>({
     },
     laptopPasswordChanged: {
         type: Date,
+        required: false,
+    },
+    autoOrder: [
+        {
+            chosen: {
+                type: String,
+                required: false,
+            },
+        },
+    ],
+    blocked: {
+        type: Boolean,
         required: false,
     },
 });